chore: Move auth DTOs to auth directory
